pyproject.toml
[tool.poetry]
name = "aws-open-data-registry-neural-search"
version = "0.1.0"
description = ""
authors = ["Gregory Lindsey <gclindsey@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "aws_open_data_registry_neural_search"}]

[tool.poetry.dependencies]
python = "^3.11"
weaviate-client = "^3.9.0"
pyyaml = "^6.0"
boto3 = "^1.26.21"
pandas = "^1.5.2"
jmespath = "^1.0.1"
streamlit = "^1.16.0"
python-dotenv = "^0.21.0"
streamlit-ace = "^0.1.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.17.1"
jupyterthemes = "^0.20.0"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
